refactor: move signing specific code from utils.h to signing.h
